fix: ignore underscore (_) characters when calculating the missing `operationId`
When a spec doesn't specify an `operationId` for an endpoint, we need to calculate one ourselves.

This is because we need a unique name for the TS type.

We calculate this ID from the endpoint's method and path.
However, we've noticed that, if the path contains an underscore, we later cannot build the `Config` object for that endpoint.
The solution is to ignore underscores when calculating the `operationId`